fix 지하철 경로 구하기 요구사항 반영 및 line, JGraph 의존성 분리

+++ src/main/java/nextstep/subway/domain/Path.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public class Path {
     private List<Station> stations;
-    private int distance;
+    private double distance;
 
-    public Path(List<Station> stations, int distance) {
+    public Path(List<Station> stations, double distance) {
         this.stations = stations;
         this.distance = distance;
     }
@@ -16,6 +16,6 @@ public class Path {
     }
 
     public int getDistance() {
-        return distance;
+        return (int) distance;
     }
 }
